corrected some bugs

+++ push_swap_v2/SOURCES/libft/ft_lstmean.c
@@ -6,13 +6,13 @@
 /*   By: mvenanci@student.42lisboa.com <mvenanci    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 14:13:34 by mvenanci          #+#    #+#             */
-/*   Updated: 2022/11/24 18:31:02 by mvenanci@st      ###   ########.fr       */
+/*   Updated: 2022/11/28 18:12:52 by mvenanci@st      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	lmean(t_list *lst)
+long int	lmean(t_list *lst)
 {
 	return ((lsum(lst) / lstsize(lst)));
 }
