+++ push_swap_v2/SOURCES/libft/ft_lstsize.c
@@ -6,13 +6,13 @@
 /*   By: mvenanci@student.42lisboa.com <mvenanci    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/15 10:48:45 by mvenanci          #+#    #+#             */
-/*   Updated: 2022/11/20 14:09:09 by mvenanci@st      ###   ########.fr       */
+/*   Updated: 2022/11/28 17:59:23 by mvenanci@st      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	lstsize(t_list *lst)
+long int	lstsize(t_list *lst)
 {
 	int	i;
