+++ push_swap_v2/SOURCES/libft/ft_lstsum.c
@@ -6,15 +6,15 @@
 /*   By: mvenanci@student.42lisboa.com <mvenanci    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 13:59:52 by mvenanci          #+#    #+#             */
-/*   Updated: 2022/11/24 18:30:49 by mvenanci@st      ###   ########.fr       */
+/*   Updated: 2022/11/28 17:58:37 by mvenanci@st      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	lsum(t_list *lst)
+long int	lsum(t_list *lst)
 {
-	int		sum;
+	long int		sum;
 
 	if (lst)
 	{
